Extract reply presentation helpers
